add ebuilds dropped from the ixit overlay
We have dropped enlightenment from the ixit overlay because I am no
longer maintaining it because I'm no longer using the software and
I am not going to in the near future.

Fortunately active users of Gentoo and Enlightenment expressed
interest in maintaining the ebuilds and that is how this new overlay
has been started.

 * Import all enlightenment related ebuilds from a snapshot of the
   ixit overlay before they were removed.
 * Import release 0.20.2 ebuild from a pull request created for the
   ixit overlay.
 * Import package sets for enlightenment from the ixit overlay.

Source: https://github.com/okias/ixit
Closes: https://github.com/okias/ixit/pull/15